UNET/requirements.txt
Flask==2.0.1
matplotlib==3.5.3
numpy==1.21.6
Pillow==10.0.1
torch==1.8.1
torchvision==0.9.1
